refactor(utils): merged common with util
